refactor: move contract creation query to ape_ethereum
